backend/requirements.txt
pyvi
ftfy
tqdm
regex
numpy
rapidfuzz
matplotlib
opencv-python
langdetect==1.0.9
googletrans==3.1.0a0
sentence_transformers
transformers
scikit-learn
open_clip_torch
spacy